[Notebook] Feedback Updates
